requirements.txt
yt-dlp
google-api-python-client
google-auth-oauthlib
google-auth-httplib2

